modules/simulations/sagemaker-managed/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.20.0
    # via
    #   -r requirements.in
    #   cdk-nag
boto3==1.24.34
    # via -r requirements.in
botocore==1.27.34
    # via
    #   boto3
    #   s3transfer
cattrs==22.1.0
    # via jsii
cdk-nag==2.12.29
    # via -r requirements.in
constructs==10.1.52
    # via
    #   aws-cdk-lib
    #   cdk-nag
exceptiongroup==1.0.0rc8
    # via cattrs
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.62.0
    # via
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
typing-extensions==4.3.0
    # via jsii
urllib3==1.26.10
    # via botocore
